+++ test/test_context_selector.py
@@ -1,6 +1,6 @@
 from typing import Any, Dict
 
-import unittest
+import pytest
 
 from carl.envs.gymnasium.classic_control.carl_pendulum import CARLPendulum
 from carl.utils.types import Context
@@ -8,7 +8,7 @@ from carl.utils.types import Context
 CARLPendulum.render_mode = "rgb_array"
 
 
-class TestContextSelection(unittest.TestCase):
+class TestContextSelection:
     @staticmethod
     def generate_contexts() -> Dict[Any, Context]:
         keys = "abc"
@@ -23,11 +23,11 @@ class TestContextSelection(unittest.TestCase):
         env = CARLPendulum(contexts=contexts)
 
         env.reset()
-        self.assertEqual(type(env.context_selector), RoundRobinSelector)
-        self.assertEqual(env.context_selector.n_calls, 1)
+        assert isinstance(env.context_selector, RoundRobinSelector)
+        assert env.context_selector.n_calls == 1
 
         env.reset()
-        self.assertEqual(env.context_selector.n_calls, 2)
+        assert env.context_selector.n_calls == 2
 
     def test_roundrobin_selector_init(self):
         from carl.context.selection import RoundRobinSelector
@@ -45,16 +45,16 @@ class TestContextSelection(unittest.TestCase):
         env = CARLPendulum(
             contexts=contexts, context_selector=RandomSelector(contexts=contexts)
         )
-        self.assertEqual(type(env.context_selector), RandomSelector)
+        assert isinstance(env.context_selector, RandomSelector)
 
     def test_random_selectorclass_init(self):
         from carl.context.selection import RandomSelector
 
         contexts = self.generate_contexts()
         env = CARLPendulum(contexts=contexts, context_selector=RandomSelector)
-        self.assertEqual(type(env.context_selector), RandomSelector)
+        assert isinstance(env.context_selector, RandomSelector)
 
     def test_unknown_selector_init(self):
-        with self.assertRaises(ValueError):
+        with pytest.raises(ValueError):
             contexts = self.generate_contexts()
             _ = CARLPendulum(contexts=contexts, context_selector="bork")